feat : Response 수정 외 1
<description>
- Response 필드 수정
- 테스트 코드 작성 중

+++ src/main/kotlin/me/hanhyur/kopring/macallan/product/entity/ProductOption.kt
@@ -19,6 +19,6 @@ class ProductOption(
     var quantity: Int,
 
     @Column(name = "discount")
-    var discount: Double,
+    var discount: Int,
 ): BaseEntity() {
 }
